Can draw a box via directx 12.

+++ src/render/render_api/fence.cpp
@@ -12,12 +12,12 @@ Fence::~Fence()
 	close_event_handle(fence_handle);
 }
 
-void Fence::create(Gpu_Device &device, u32 initial_value)
+void Fence::create(Gpu_Device &device, u64 initial_value)
 {
 	HR(device->CreateFence(initial_value, D3D12_FENCE_FLAG_NONE, IID_PPV_ARGS(release_and_get_address())));
 }
 
-void Fence::wait_for_gpu(u32 fence_value)
+void Fence::wait_for_gpu(u64 fence_value)
 {
 	if (d3d12_object->GetCompletedValue() < fence_value) {
 		d3d12_object->SetEventOnCompletion(fence_value, fence_handle);
